+++ PoshBot/Classes/Connection.ps1
@@ -1,7 +1,7 @@
 
 # This class represents the connection to a backend Chat network
 
-class Connection {
+class Connection : BaseLogger {
     [ConnectionConfig]$Config
     [ConnectionStatus]$Status = [ConnectionStatus]::Disconnected