+++ PoshBot/Classes/ExceptionFormatter.ps1
@@ -1,13 +1,18 @@
 
 class ExceptionFormatter {
+
+    static [pscustomobject]Summarize([System.Management.Automation.ErrorRecord]$Exception) {
+        return [pscustomobject]@{
+            CommandName = $Exception.InvocationInfo.MyCommand.Name
+            Message = $Exception.Exception.Message
+            TargetObject = $Exception.TargetObject
+            Position = $Exception.InvocationInfo.PositionMessage
+            CategoryInfo = $Exception.CategoryInfo.ToString()
+            FullyQualifiedErrorId = $Exception.FullyQualifiedErrorId
+        }
+    }
+
     static [string]ToJson([System.Management.Automation.ErrorRecord]$Exception) {
-        $props = @(
-            @{l = 'CommandName'; e = {$_.InvocationInfo.MyCommand.Name}}
-            @{l = 'Message'; e = {$_.Exception.Message}}
-            @{l = 'Position'; e = {$_.InvocationInfo.PositionMessage}}
-            @{l = 'CategoryInfo'; e = {$_.CategoryInfo.ToString()}}
-            @{l = 'FullyQualifiedErrorId'; e = {$_.FullyQualifiedErrorId}}
-        )
-        return $Exception | Select-Object -Property $props | ConvertTo-Json
+        return ([ExceptionFormatter]::Summarize($Exception) | ConvertTo-Json)
     }
 }