+++ PoshBot/Classes/Logger.ps1
@@ -1,75 +1,4 @@
 
-enum LogSeverity {
-    Normal
-    Warning
-    Error
-}
-
-class LogMessage {
-    [datetime]$DateTime = (Get-Date)
-    [string]$Severity = [LogSeverity]::Normal
-    [string]$LogLevel = [LogLevel]::Info
-    [string]$Message
-    [object]$Data
-
-    LogMessage() {
-    }
-
-    LogMessage([string]$Message) {
-        $this.Message = $Message
-    }
-
-    LogMessage([string]$Message, [object]$Data) {
-        $this.Message = $Message
-        $this.Data = $Data
-    }
-
-    LogMessage([LogSeverity]$Severity, [string]$Message) {
-        $this.Severity = $Severity
-        $this.Message = $Message
-    }
-
-    LogMessage([LogSeverity]$Severity, [string]$Message, [object]$Data) {
-        $this.Severity = $Severity
-        $this.Message = $Message
-        $this.Data = $Data
-    }
-
-    # Borrowed from https://github.com/PowerShell/PowerShell/issues/2736
-    hidden [string]Compact([string]$Json) {
-        $indent = 0
-        $compacted = ($Json -Split '\n' | ForEach-Object {
-            if ($_ -match '[\}\]]') {
-                # This line contains  ] or }, decrement the indentation level
-                $indent--
-            }
-            $line = (' ' * $indent * 2) + $_.TrimStart().Replace(':  ', ': ')
-            if ($_ -match '[\{\[]') {
-                # This line contains [ or {, increment the indentation level
-                $indent++
-            }
-            $line
-        }) -Join "`n"
-        return $compacted
-    }
-
-    [string]ToJson() {
-        $json = @{
-            DataTime = $this.DateTime
-            Severity = $this.Severity
-            LogLevel = $this.LogLevel
-            Message = $this.Message
-            Data = $this.Data
-        } | ConvertTo-Json -Depth 100 -Compress
-        return $json
-        #return $this.Compact($json)
-    }
-
-    [string]ToString() {
-        return $this.ToJson()
-    }
-}
-
 class Logger {
 
     # The log directory
@@ -77,22 +6,25 @@ class Logger {
 
     hidden [string]$LogFile
 
-    # Out logging level
+    # Our logging level
     # Any log messages less than or equal to this will be logged
     [LogLevel]$LogLevel
 
     # The max size for the log files before rolling
-    [int]$MaxSizeMB = 10
+    [int]$MaxSizeMB
 
     # Number of each log file type to keep
-    [int]$FilesToKeep = 5
+    [int]$FilesToKeep
 
     # Create logs files under provided directory
-    Logger([string]$LogDir, [LogLevel]$LogLevel) {
+    Logger([string]$LogDir, [LogLevel]$LogLevel, [int]$MaxLogSizeMB, [int]$MaxLogsToKeep) {
         $this.LogDir = $LogDir
         $this.LogLevel = $LogLevel
+        $this.MaxSizeMB = $MaxLogSizeMB
+        $this.FilesToKeep = $MaxLogsToKeep
         $this.LogFile = Join-Path -Path $this.LogDir -ChildPath 'PoshBot.log'
         $this.CreateLogFile()
+        $this.Log([LogMessage]::new("Log level set to [$($this.LogLevel)]"))
     }
 
     # Create new log file or roll old log
@@ -104,55 +36,78 @@ class Logger {
         New-Item -Path $this.LogFile -ItemType File -Force
     }
 
-    [void]Info([LogMessage]$Message) {
-        $Message.LogLevel = [LogLevel]::Info
-        $this.Log($Message)
-    }
-
-    [void]Verbose([LogMessage]$Message) {
-        $Message.LogLevel = [LogLevel]::Verbose
-        $this.Log($Message)
-    }
-
-    [void]Debug([LogMessage]$Message) {
-        $Message.LogLevel = [LogLevel]::Debug
-        $this.Log($Message)
-    }
-
-    # Write out message in JSON form to log file
+    # Log the message and optionally write to console
     [void]Log([LogMessage]$Message) {
-
-        if ($global:VerbosePreference -eq 'Continue') {
-            Write-Verbose -Message $Message.ToJson()
-        } elseIf ($global:DebugPreference -eq 'Continue') {
-            Write-Debug -Message $Message.ToJson()
+        switch ($Message.Severity.ToString()) {
+            'Normal' {
+                if ($global:VerbosePreference -eq 'Continue') {
+                    Write-Verbose -Message $Message.ToJson()
+                } elseIf ($global:DebugPreference -eq 'Continue') {
+                    Write-Debug -Message $Message.ToJson()
+                }
+                break
+            }
+            'Warning' {
+                if ($global:WarningPreference -eq 'Continue') {
+                    Write-Warning -Message $Message.ToJson()
+                }
+                break
+            }
+            'Error' {
+                if ($global:ErrorActionPreference -eq 'Continue') {
+                    Write-Error -Message $Message.ToJson()
+                }
+                break
+            }
         }
 
-        if ($Message.LogLevel.value__ -le $This.LogLevel.value__) {
+        if ($Message.LogLevel.value__ -le $this.LogLevel.value__) {
             $this.RollLog($this.LogFile, $false)
             $json = $Message.ToJson()
-            $json | Out-File -FilePath $this.LogFile -Append -Encoding utf8
+            $this.WriteLine($json)
         }
+    }
+
+    [void]Log([LogMessage]$Message, [string]$LogFile, [int]$MaxLogSizeMB, [int]$MaxLogsToKeep) {
+        $this.RollLog($LogFile, $false, $MaxLogSizeMB, $MaxLogSizeMB)
+        $json = $Message.ToJson()
+        $sw = [System.IO.StreamWriter]::new($LogFile, [System.Text.Encoding]::UTF8)
+        $sw.WriteLine($json)
+        $sw.Close()
+    }
+
+    # Write line to file
+    hidden [void]WriteLine([string]$Message) {
+        $sw = [System.IO.StreamWriter]::new($this.LogFile, [System.Text.Encoding]::UTF8)
+        $sw.WriteLine($Message)
+        $sw.Close()
+    }
+
+    hidden [void]RollLog([string]$LogFile, [bool]$Always) {
+        $this.RollLog($LogFile, $Always, $this.MaxSizeMB, $this.FilesToKeep)
     }
 
     # Checks to see if file in question is larger than the max size specified for the logger.
     # If it is, it will roll the log and delete older logs to keep our number of logs per log type to
     # our max specifiex in the logger.
     # Specified $Always = $true will roll the log regardless
-    hidden [void]RollLog([string]$LogFile, [bool]$Always) {
+    hidden [void]RollLog([string]$LogFile, [bool]$Always, $MaxLogSize, $MaxFilesToKeep) {
+
+        $keep = $MaxFilesToKeep - 1
+
         if (Test-Path -Path $LogFile) {
-            if ((($file = Get-Item -Path $logFile) -and ($file.Length/1mb) -gt $this.MaxSizeMB) -or $Always) {
+            if ((($file = Get-Item -Path $logFile) -and ($file.Length/1mb) -gt $MaxLogSize) -or $Always) {
                 # Remove the last item if it would go over the limit
-                if (Test-Path -Path "$logFile.$($this.FilesToKeep)") {
-                    Remove-Item -Path "$logFile.$($this.FilesToKeep)"
+                if (Test-Path -Path "$logFile.$keep") {
+                    Remove-Item -Path "$logFile.$keep"
                 }
-                foreach ($i in $($this.FilesToKeep)..1) {
+                foreach ($i in $keep..1) {
                     if (Test-path -Path "$logFile.$($i-1)") {
                         Move-Item -Path "$logFile.$($i-1)" -Destination "$logFile.$i"
                     }
                 }
                 Move-Item -Path $logFile -Destination "$logFile.$i"
-                $null = New-Item -Path $LogFile -Type File -Force | Out-Null
+                New-Item -Path $LogFile -Type File -Force > $null
             }
         }
     }