+++ PoshBot/Public/New-PoshBotConfiguration.ps1
@@ -22,6 +22,16 @@ function New-PoshBotConfiguration {
         loaded when PoshBot starts.
     .PARAMETER LogLevel
         The level of logging that PoshBot will do.
+    .PARAMETER MaxLogSizeMB
+        The maximum log file size in megabytes.
+    .PARAMETER MaxLogsToKeep
+        The maximum number of logs to keep. Once this value is reached, logs will start rotating.
+    .PARAMETER LogCommandHistory
+        Enable command history to be logged to a separate file for convenience. The default it $true
+    .PARAMETER CommandHistoryMaxLogSizeMB
+        The maximum log file size for the command history
+    .PARAMETER CommandHistoryMaxLogsToKeep
+        The maximum number of logs to keep for command history. Once this value is reached, the logs will start rotating.
     .PARAMETER BackendConfiguration
         A hashtable of configuration options required by the backend chat network implementation.
     .PARAMETER PluginConfiguration
@@ -161,6 +171,11 @@ function New-PoshBotConfiguration {
         [string[]]$PluginRepository = @('PSGallery'),
         [string[]]$ModuleManifestsToLoad = @(),
         [LogLevel]$LogLevel = [LogLevel]::Verbose,
+        [int]$MaxLogSizeMB = 10,
+        [int]$MaxLogsToKeep = 5,
+        [bool]$LogCommandHistory = $true,
+        [int]$CommandHistoryMaxLogSizeMB = 10,
+        [int]$CommandHistoryMaxLogsToKeep = 5,
         [hashtable]$BackendConfiguration = @{},
         [hashtable]$PluginConfiguration = @{},
         [string[]]$BotAdmins = @(),
@@ -182,6 +197,11 @@ function New-PoshBotConfiguration {
     $config.CommandPrefix = $CommandPrefix
     $config.LogDirectory = $LogDirectory
     $config.LogLevel = $LogLevel
+    $config.MaxLogSizeMB = $MaxLogSizeMB
+    $config.MaxLogsToKeep = $MaxLogsToKeep
+    $config.LogCommandHistory = $LogCommandHistory
+    $config.CommandHistoryMaxLogSizeMB = $CommandHistoryMaxLogSizeMB
+    $config.CommandHistoryMaxLogsToKeep = $CommandHistoryMaxLogsToKeep
     $config.BackendConfiguration = $BackendConfiguration
     $config.PluginConfiguration = $PluginConfiguration
     $config.ModuleManifestsToLoad = $ModuleManifestsToLoad